+++ GcproExtensionApp/LL/EL/FormScaleAdapter.cs
@@ -1256,7 +1256,7 @@ namespace GcproExtensionApp
             if (ComboDiagram.SelectedItem != null)
             {
                 selectedDiagram = ComboDiagram.SelectedItem.ToString();
-                objScaleAdapter.Diagram = ScaleAdapter.ParseInfoValue(selectedDiagram, AppGlobal.FIELDS_SEPARATOR, AppGlobal.NO_DIAGRAM);
+                objScaleAdapter.Diagram = (int)ScaleAdapter.ParseInfoValue(selectedDiagram, AppGlobal.FIELDS_SEPARATOR, AppGlobal.NO_DIAGRAM);
             }
             ///<Page></Page>
             objScaleAdapter.Page = txtPage.Text;
