+++ GcproExtensionApp/LL/MA/FormMA_Roll8Stand.cs
@@ -1627,7 +1627,7 @@ namespace GcproExtensionApp
             if (ComboDiagram.SelectedItem != null)
             {
                 selectedDiagram = ComboDiagram.SelectedItem.ToString();
-                myRoll8Stand.Diagram = Roll8Stand.ParseInfoValue(selectedDiagram, AppGlobal.FIELDS_SEPARATOR, AppGlobal.NO_DIAGRAM);
+                myRoll8Stand.Diagram = (int)Roll8Stand.ParseInfoValue(selectedDiagram, AppGlobal.FIELDS_SEPARATOR, AppGlobal.NO_DIAGRAM);
             }
             ///<Page></Page>
             myRoll8Stand.Page = txtPage.Text;
